Move Aceternity UI components to its own folder
